--- README copy.md
@@ -1,1 +1,0 @@
-# react-gift-order
