+++ README.md
@@ -1,9 +1,1 @@
-# react-gift-login
-
-홈과 로그인(WIP) 레포지토리를 클론.
-불필요한 파일, 코드를 정리.
-절대경로로 Import 할 수 있게 alias를 설정.
-Prettier가 무엇인지 찾아보고 추가.
-emotion 스타일 라이브러리를 추가하고, reset css를 적용.
-기본 폰트로 Pretendard를 적용.
-기능 단위로 나누어 커밋.
+# react-gift-order
