+++ src/Components/Pages/Home.tsx
@@ -1,11 +1,11 @@
-import Layout from '@/Components/Layout/Layout';
-import NavigationBar from '@/Components/NavigationBar/NavigationBar';
-import SelectFriend from '@/Components/Section/SelectFriend';
-import GiftCategoryList from '@/Components/Category/GiftCategoryList';
-import Banner from '@/Components/Section/Banner';
-import GiftGrid from '@/Components/GiftRanking/GiftGrid';
-import GiftItem from '@/Components/GiftRanking/GiftItem';
-import Filter from '@/Components/Filter/Filter';
+import Layout from '@/components/layout/Layout';
+import NavigationBar from '@/components/navigation-bar/NavigationBar';
+import SelectFriend from '@/components/section/SelectFriend';
+import GiftCategoryList from '@/components/category/GiftCategoryList';
+import Banner from '@/components/section/Banner';
+import GiftGrid from '@/components/gift-ranking/GiftGrid';
+import GiftItem from '@/components/gift-ranking/GiftItem';
+import Filter from '@/components/filter/Filter';
 
 function Home() {
   return (
